adjusted most of the is[hand type] funcs and the highCard func
the aces are represented as value 1, for easy use with the reader, yet they are the strongest card. so a special case for all those funcs has been added where if cards value is 1, instead of returning the value it returns 13.5 (more than 13 for kings). Also adjusted rateHand to multiply by 15 to avoid troubles with the new 13.5 value